Very Ascending Sequence Checker
📈 Very Ascending Sequence Checker
Description:
This C program takes a sequence of user-input integers and checks whether the entire sequence is strictly increasing (each number is greater than the previous one). If it is, the sequence is labeled as Very Ascending. 🔢⬆️

Functions & Concepts Used:
printf() and scanf() for input/output 🖨️
do-while loop for iterative input 🔁
Conditional statements (if-else) for logical checks ✅
Variable tracking of current and previous values to compare ➕➖